backend/pyproject.toml
[project]
name = "carrierassure"
version = "0.1.0"
description = "Carrier Test"
authors = [
    {name = "Hector"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django (>=6.0.3,<7.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "pymongo (>=4.7,<5.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-mongodb-backend (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.4,<0.16.0)",
    "black (>=26.1.0,<27.0.0)"
]

[tool.black]
line-length = 100
target-version = ["py312"]
exclude = '''
/(
  migrations
)/
'''
